server/requirements.txt
# server/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.27
python-magic==0.4.27
Werkzeug==3.0.1